Introduce ekaf_max_downtime_buffer_size option
When kafka downtime is detected, we store messages in-memory
Until now, this was unbounded, assuming kafka restarts are fast.

Optionally, you can now bound this max queue.

    %% eg:
    application:set_env(ekaf, ekaf_max_downtime_buffer_size, 10000000).

This is a topic level queue, so set this based on the memory on the
machine in question, anticipated downtime, etc

Updated readme with tests