Add ?EKAF_CALLBACK_MASSAGE_BUFFER, encode_messages_as_one_large_json to send batch as 1 msg
Usually in ekaf’s batching behaviour - we buffer for 5 seconds or until the topic max timeout
(both configurable) and then send as many messages as buffered in 1 large message set

Samza has a nice functionality where consumers can read in batches.
This commit adds similar behaviour by doing the batching on the producer side.

= EARLIER =

e.g.: earlier if you call

    Event1 = [{<<“id”>>, 1}]
    ekaf:produce_async_batched(<<“events”>>, jsx:encode(Event1)}

    Event2 = [{<<“id”>>, 2}]
    ekaf:produce_async_batched(<<“events”>>, jsx:encode(Event2)}

it used to be sent as two messages as part of 1 messageset
because it calls

    ekaf_lib:data_to_message_sets(Messages).

and hence your consumer would get two messages

    # on kafka console consumer you will see
    {“id”:1}
    {“id”:2}

= THE CHANGE =

With this commit, if you set

    application:set_env(ekaf,
                        ?EKAF_CALLBACK_MASSAGE_BUFFER,
                        %% this fn is just two lines and calls jx
                        %% feel free to move to your callback instead
                        {ekaf_callbacks, encode_messages_as_one_large_json}),

Now sending events like

    Event3 = [{<<“id”>>, 1}]
    ekaf:produce_async_batched(<<“events”>>, [Event3]}

    Event4 = [{<<“id”>>, 2}]
    %% NOTE: we're not sending a list of proplists. now an encoded binary json
    ekaf:produce_async_batched(<<“events”>>, [Event4]}

it will call the ?EKAF_CALLBACK_MASSAGE_BUFFER callback set above which does

   ekaf_lib:data_to_message_sets(jsx:encode(Messages)).

Now the messageset will only have 1 message, which is the current
buffer all encoded as json. Hence the consumer will now get

    # on kafka console consumer you will see
    [{“id”:1}, {“id”:2}]

1) set the ?EKAF_CALLBACK_MASSAGE_BUFFER
2) do not encode each individual message, instead
   do it once for each buffer in ^ callback

= AREAS OF USE =

You can push debug logs into a topic with just few workers
The consumer will get multiple debug events in 1 message
Which it can just format and email as a digest of errors

This again is inspired by Samza batching, but implemented
on the producer instead of the state/collection/batching
happening on the consumer